Assessment/question4/requirements.txt
django
djangorestframework
dj-rest-auth